benches: Re-organize module layout
Code in `mod.rs` of `aarch64` do not contain architecture specific code
as of `aarch64`, it's only used by `aarch64` to run benchmark tests,
move it one level up and rename it to `fdt.rs` to allow reusing on
riscv64 platform.
